requirements.txt
Django==1.11.7
django-annoying==0.10.3
django-crispy-forms==1.7.0
gunicorn==19.7.1
pytz==2017.3
